docker-cc-uas/consumer-url/requirements.txt
mysql-connector-python
python-logstash
scrapy
bs4
flask
elasticsearch
logstash
flask_cors
json-logging-py
pymongo
Werkzeug
pika
requests